Simple solution to get_map_center issues
I had 10 different functions to get the center of the map and it still
was getting it wrong in some cases. I realize the problem was in the
idea of doing that by accessing the central node of the linked list.
- Now I need one simple function get_map_center, based on a max_x - min_x logic
- I just had to create get_xy_range and call it with get_altitude_range
  right before the map->get_center(d) function call in the render_map_bonus.
- only when not spherical... for the spherical I only calculate it once
  and not in the draw loop. It works but I dont know why yet.
-->> PROBLEM TO FIX: new change_altitude methods might be pretty, but
they don't work! How could I commit them before testing???