All Clean
There's now a clear rule for when functions should be defined in a
header or not.
-> When there are excess function definitions in a source file, I can
extract the simple ones and if they deal exclusively with the attributes
of a given "class", then I move their definition to the header file of
that "class". Or at least, almost exclusively.
-> Methods are now only called from within an "object", this way
I can always know that it is a method and is defined in a header and not
a common function defined elsewhere
-> benefits of this practice:
1 -) I need to follow a 42 strict rule of never defining more than five
functions in a file, so this practice avoids having to create source
files with weird names such as set_hex_color_aux or compute_color_2
2 -) Hide some implementation details making a source file more readable